Ex 6.8: anecdotes, step 6. Separated rendering of anecdote list into component <AnecdoteList />. Moved all logic related to voting for an anecdote into this component. Modified App component accordingly.
