+++ submissions/anecdotes-redux/src/App.js
@@ -1,30 +1,13 @@
 import React from 'react'
 import AnecdoteForm from './components/AnecdoteForm'
-import { useSelector, useDispatch } from 'react-redux'
-import { voteForAnecdote, createAnecdote } from './reducers/anecdoteReducer'
+import AnecdoteList from './components/AnecdoteList'
 
 const App = () => {
-  const anecdotes = useSelector(state => state)
-  const dispatch = useDispatch()
-
-  const vote = (id) => dispatch(voteForAnecdote(id))
-
   return (
     <div>
       <h2>Anecdotes</h2>
-      {anecdotes.map(anecdote =>
-        <div key={anecdote.id}>
-          <div>
-            {anecdote.content}
-          </div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      )}
-      <AnecdoteForm dispatch={dispatch} />
-
+      <AnecdoteList />
+      <AnecdoteForm />
     </div>
   )
 }
