+++ submissions/anecdotes-redux/src/components/AnecdoteForm.js
@@ -1,7 +1,9 @@
 import React from 'react'
+import { useDispatch } from 'react-redux'
 import { createAnecdote } from '../reducers/anecdoteReducer'
 
-const AnecdoteForm = ( { dispatch } ) => {
+const AnecdoteForm = () => {
+    const dispatch = useDispatch()
 
     const addAnecdote = event => {
         event.preventDefault()
